+++ android/contrib/tools/do-compile-ffmpeg.sh
@@ -25,53 +25,22 @@ echo "[*] check env $1"
 echo "===================="
 set -e
 
-UNAME_S=$(uname -s)
-UNAME_SM=$(uname -sm)
-echo "build on $UNAME_SM"
-
-echo "ANDROID_SDK=$ANDROID_SDK"
-echo "ANDROID_NDK=$ANDROID_NDK"
-
-if [ -z "$ANDROID_NDK" -o -z "$ANDROID_SDK" ]; then
-    echo "You must define ANDROID_NDK, ANDROID_SDK before starting."
-    echo "They must point to your NDK and SDK directories."
-    echo ""
-    exit 1
-fi
 
 #--------------------
 # common defines
 FF_ARCH=$1
+FF_BUILD_OPT=$2
+echo "FF_ARCH=$FF_ARCH"
+echo "FF_BUILD_OPT=$FF_BUILD_OPT"
 if [ -z "$FF_ARCH" ]; then
     echo "You must specific an architecture 'arm, armv7a, x86, ...'."
     echo ""
     exit 1
 fi
 
-# try to detect NDK version
-FF_NDK_REL=$(grep -o '^r[0-9]*.*' $ANDROID_NDK/RELEASE.TXT 2>/dev/null|cut -b2-)
-case "$FF_NDK_REL" in
-    9*|10*)
-        # we don't use 4.4.3 because it doesn't handle threads correctly.
-        if test -d ${ANDROID_NDK}/toolchains/arm-linux-androideabi-4.8
-        # if gcc 4.8 is present, it's there for all the archs (x86, mips, arm)
-        then
-            echo "NDKr$FF_NDK_REL detected"
-        else
-            echo "You need the NDKr9 or later"
-            exit 1
-        fi
-    ;;
-    *)
-        echo "You need the NDKr9 or later"
-        exit 1
-    ;;
-esac
 
 FF_BUILD_ROOT=`pwd`
 FF_ANDROID_PLATFORM=android-9
-FF_GCC_VER=4.8
-FF_GCC_64_VER=4.9
 
 
 FF_BUILD_NAME=
@@ -89,26 +58,18 @@ FF_DEP_LIBS=
 FF_MODULE_DIRS="compat libavcodec libavfilter libavformat libavutil libswresample libswscale"
 FF_ASSEMBLER_SUB_DIRS=
 
+
 #--------------------
 echo ""
 echo "--------------------"
 echo "[*] make NDK standalone toolchain"
 echo "--------------------"
-FF_MAKE_TOOLCHAIN_FLAGS=
-case "$UNAME_S" in
-    Darwin)
-        FF_MAKE_TOOLCHAIN_FLAGS="$FF_MAKE_TOOLCHAIN_FLAGS --system=darwin-x86_64"
-    ;;
-    CYGWIN_NT-*)
-        FF_MAKE_TOOLCHAIN_FLAGS="$FF_MAKE_TOOLCHAIN_FLAGS --system=windows-x86_64"
+. ./tools/do-detect-env.sh
+FF_MAKE_TOOLCHAIN_FLAGS=$IJK_MAKE_TOOLCHAIN_FLAGS
+FF_MAKE_FLAGS=$IJK_MAKE_FLAG
+FF_GCC_VER=$IJK_GCC_VER
+FF_GCC_64_VER=$IJK_GCC_64_VER
 
-        FF_WIN_TEMP="$(cygpath -am /tmp)"
-        export TEMPDIR=$FF_WIN_TEMP/
-
-        echo "Cygwin temp prefix=$FF_WIN_TEMP/"
-        #FF_CFG_FLAGS="$FF_CFG_FLAGS --tempprefix=$FF_WIN_TEMP/"
-    ;;
-esac
 
 #----- armv7a begin -----
 if [ "$FF_ARCH" = "armv7a" ]; then
@@ -197,6 +158,15 @@ else
     exit 1
 fi
 
+if [ ! -d $FF_SOURCE ]; then
+    echo ""
+    echo "!! ERROR"
+    echo "!! Can not find FFmpeg directory for $FF_BUILD_NAME"
+    echo "!! Run 'sh init-android.sh' first"
+    echo ""
+    exit 1
+fi
+
 FF_TOOLCHAIN_PATH=$FF_BUILD_ROOT/build/$FF_BUILD_NAME/toolchain
 FF_MAKE_TOOLCHAIN_FLAGS="$FF_MAKE_TOOLCHAIN_FLAGS --install-dir=$FF_TOOLCHAIN_PATH"
 
@@ -212,17 +182,10 @@ case "$UNAME_S" in
     ;;
 esac
 
+
 mkdir -p $FF_PREFIX
 mkdir -p $FF_SYSROOT
 
-FF_MAKEFLAGS=
-if which nproc >/dev/null
-then
-    FF_MAKEFLAGS=-j`nproc`
-elif [ "$UNAME_S" = "Darwin" ] && which sysctl >/dev/null
-then
-    FF_MAKEFLAGS=-j`sysctl -n machdep.cpu.thread_count`
-fi
 
 FF_TOOLCHAIN_TOUCH="$FF_TOOLCHAIN_PATH/touch"
 if [ ! -f "$FF_TOOLCHAIN_TOUCH" ]; then
@@ -269,6 +232,7 @@ export COMMON_FF_CFG_FLAGS=
 #--------------------
 # with openssl
 if [ -f "${FF_DEP_OPENSSL_LIB}/libssl.a" ]; then
+    echo "OpenSSL detected"
 # FF_CFG_FLAGS="$FF_CFG_FLAGS --enable-nonfree"
     FF_CFG_FLAGS="$FF_CFG_FLAGS --enable-openssl"
 
@@ -294,6 +258,19 @@ FF_CFG_FLAGS="$FF_CFG_FLAGS --enable-pic"
 FF_CFG_FLAGS="$FF_CFG_FLAGS --enable-asm"
 FF_CFG_FLAGS="$FF_CFG_FLAGS --enable-inline-asm"
 
+case "$FF_BUILD_OPT" in
+    debug)
+        FF_CFG_FLAGS="$FF_CFG_FLAGS --disable-optimizations"
+        FF_CFG_FLAGS="$FF_CFG_FLAGS --enable-debug"
+        FF_CFG_FLAGS="$FF_CFG_FLAGS --disable-small"
+    ;;
+    *)
+        FF_CFG_FLAGS="$FF_CFG_FLAGS --enable-optimizations"
+        FF_CFG_FLAGS="$FF_CFG_FLAGS --enable-debug"
+        FF_CFG_FLAGS="$FF_CFG_FLAGS --enable-small"
+    ;;
+esac
+
 #--------------------
 echo ""
 echo "--------------------"
@@ -316,7 +293,7 @@ echo "--------------------"
 echo "[*] compile ffmpeg"
 echo "--------------------"
 cp config.* $FF_PREFIX
-make $FF_MAKEFLAGS
+make $FF_MAKE_FLAGS > /dev/null
 make install
 mkdir -p $FF_PREFIX/include/libffmpeg
 cp -f config.h $FF_PREFIX/include/libffmpeg/config.h