+++ android/contrib/tools/do-compile-openssl.sh
@@ -17,14 +17,11 @@
 #
 
 #--------------------
-echo "===================="
-echo "[*] check env $1"
-echo "===================="
 set -e
 
-if [ -z "$ANDROID_NDK" -o -z "$ANDROID_SDK" ]; then
-    echo "You must define ANDROID_NDK, ANDROID_SDK before starting."
-    echo "They must point to your NDK and SDK directories.\n"
+if [ -z "$ANDROID_NDK" ]; then
+    echo "You must define ANDROID_NDK before starting."
+    echo "They must point to your NDK directories.\n"
     exit 1
 fi
 
@@ -36,30 +33,9 @@ if [ -z "$FF_ARCH" ]; then
     exit 1
 fi
 
-# try to detect NDK version
-FF_NDK_REL=$(grep -o '^r[0-9]*.*' $ANDROID_NDK/RELEASE.TXT 2>/dev/null|cut -b2-)
-case "$FF_NDK_REL" in
-    9*|10*)
-        # we don't use 4.4.3 because it doesn't handle threads correctly.
-        if test -d ${ANDROID_NDK}/toolchains/arm-linux-androideabi-4.8
-        # if gcc 4.8 is present, it's there for all the archs (x86, mips, arm)
-        then
-            echo "NDKr$FF_NDK_REL detected"
-        else
-            echo "You need the NDKr9 or later"
-            exit 1
-        fi
-    ;;
-    7|8|*)
-        echo "You need the NDKr9 or later"
-        exit 1
-    ;;
-esac
 
 FF_BUILD_ROOT=`pwd`
 FF_ANDROID_PLATFORM=android-9
-FF_GCC_VER=4.8
-
 
 
 FF_BUILD_NAME=
@@ -71,6 +47,20 @@ FF_PLATFORM_CFG_FLAGS=
 
 FF_EXTRA_CFLAGS=
 FF_EXTRA_LDFLAGS=
+
+
+
+#--------------------
+echo ""
+echo "--------------------"
+echo "[*] make NDK standalone toolchain"
+echo "--------------------"
+. ./tools/do-detect-env.sh
+FF_MAKE_TOOLCHAIN_FLAGS=$IJK_MAKE_TOOLCHAIN_FLAGS
+FF_MAKE_FLAGS=$IJK_MAKE_FLAG
+FF_GCC_VER=$IJK_GCC_VER
+FF_GCC_64_VER=$IJK_GCC_64_VER
+
 
 #----- armv7a begin -----
 if [ "$FF_ARCH" = "armv7a" ]; then
@@ -100,6 +90,28 @@ elif [ "$FF_ARCH" = "x86" ]; then
 
     FF_PLATFORM_CFG_FLAGS="android-x86"
 
+elif [ "$FF_ARCH" = "x86_64" ]; then
+    FF_ANDROID_PLATFORM=android-21
+
+    FF_BUILD_NAME=openssl-x86_64
+    FF_SOURCE=$FF_BUILD_ROOT/$FF_BUILD_NAME
+
+    FF_CROSS_PREFIX=x86_64-linux-android
+    FF_TOOLCHAIN_NAME=${FF_CROSS_PREFIX}-${FF_GCC_64_VER}
+
+    FF_PLATFORM_CFG_FLAGS="linux-x86_64"
+
+elif [ "$FF_ARCH" = "arm64" ]; then
+    FF_ANDROID_PLATFORM=android-21
+
+    FF_BUILD_NAME=openssl-arm64
+    FF_SOURCE=$FF_BUILD_ROOT/$FF_BUILD_NAME
+
+    FF_CROSS_PREFIX=aarch64-linux-android
+    FF_TOOLCHAIN_NAME=${FF_CROSS_PREFIX}-${FF_GCC_64_VER}
+
+    FF_PLATFORM_CFG_FLAGS="linux-aarch64"
+
 else
     echo "unknown architecture $FF_ARCH";
     exit 1
@@ -113,27 +125,18 @@ FF_PREFIX=$FF_BUILD_ROOT/build/$FF_BUILD_NAME/output
 mkdir -p $FF_PREFIX
 mkdir -p $FF_SYSROOT
 
+
 #--------------------
-echo "\n--------------------"
+echo ""
+echo "--------------------"
 echo "[*] make NDK standalone toolchain"
 echo "--------------------"
-UNAME_S=$(uname -s)
-FF_MAKE_TOOLCHAIN_FLAGS="--install-dir=$FF_TOOLCHAIN_PATH"
-if [ "$UNAME_S" = "Darwin" ]; then
-    echo "build on darwin-x86_64"
-    FF_MAKE_TOOLCHAIN_FLAGS="$FF_MAKE_TOOLCHAIN_FLAGS --system=darwin-x86_64"
-    FF_MAKE_FLAG=-j`sysctl -n machdep.cpu.thread_count`
-fi
+. ./tools/do-detect-env.sh
+FF_MAKE_TOOLCHAIN_FLAGS=$IJK_MAKE_TOOLCHAIN_FLAGS
+FF_MAKE_FLAGS=$IJK_MAKE_FLAG
 
-FF_MAKEFLAGS=
-if which nproc >/dev/null
-then
-    FF_MAKEFLAGS=-j`nproc`
-elif [ "$UNAME_S" = "Darwin" ] && which sysctl >/dev/null
-then
-    FF_MAKEFLAGS=-j`sysctl -n machdep.cpu.thread_count`
-fi
 
+FF_MAKE_TOOLCHAIN_FLAGS="$FF_MAKE_TOOLCHAIN_FLAGS --install-dir=$FF_TOOLCHAIN_PATH"
 FF_TOOLCHAIN_TOUCH="$FF_TOOLCHAIN_PATH/touch"
 if [ ! -f "$FF_TOOLCHAIN_TOUCH" ]; then
     $ANDROID_NDK/build/tools/make-standalone-toolchain.sh \
@@ -143,8 +146,10 @@ if [ ! -f "$FF_TOOLCHAIN_TOUCH" ]; then
     touch $FF_TOOLCHAIN_TOUCH;
 fi
 
+
 #--------------------
-echo "\n--------------------"
+echo ""
+echo "--------------------"
 echo "[*] check openssl env"
 echo "--------------------"
 export PATH=$FF_TOOLCHAIN_PATH/bin:$PATH
@@ -162,7 +167,8 @@ FF_CFG_FLAGS="$FF_CFG_FLAGS --cross-compile-prefix=${FF_CROSS_PREFIX}-"
 FF_CFG_FLAGS="$FF_CFG_FLAGS $FF_PLATFORM_CFG_FLAGS"
 
 #--------------------
-echo "\n--------------------"
+echo ""
+echo "--------------------"
 echo "[*] configurate openssl"
 echo "--------------------"
 cd $FF_SOURCE
@@ -176,14 +182,16 @@ cd $FF_SOURCE
 #fi
 
 #--------------------
-echo "\n--------------------"
+echo ""
+echo "--------------------"
 echo "[*] compile openssl"
 echo "--------------------"
 make depend
-make
-make install
+make $FF_MAKE_FLAGS
+make install_sw
 
 #--------------------
-echo "\n--------------------"
+echo ""
+echo "--------------------"
 echo "[*] link openssl"
 echo "--------------------"