+++ ijkmedia/ijkplayer/ff_ffinc.h
@@ -31,6 +31,7 @@
 #include "libavformat/avformat.h"
 #include "libavcodec/avfft.h"
 #include "libswscale/swscale.h"
+#include "libavutil/application.h"
 #include "libavutil/base64.h"
 #include "libavutil/error.h"
 #include "libavutil/opt.h"