+++ ijkmedia/ijksdl/gles2/internal.h
@@ -72,6 +72,8 @@ typedef struct IJK_GLES2_Renderer
     int     frame_height;
     int     frame_sar_num;
     int     frame_sar_den;
+
+    GLsizei last_buffer_width;
 } IJK_GLES2_Renderer;
 
 typedef struct IJK_GLES_Matrix