+++ ijkmedia/ijksdl/gles2/renderer.c
@@ -63,13 +63,16 @@ void IJK_GLES2_Renderer_reset(IJK_GLES2_Renderer *renderer)
     if (renderer->program)
         glDeleteProgram(renderer->program);
 
+    renderer->vertex_shader   = 0;
+    renderer->fragment_shader = 0;
+    renderer->program         = 0;
+
     for (int i = 0; i < IJK_GLES2_MAX_PLANE; ++i) {
         if (renderer->plane_textures[i]) {
             glDeleteTextures(1, &renderer->plane_textures[i]);
+            renderer->plane_textures[i] = 0;
         }
     }
-
-    memset(renderer, 0, sizeof(IJK_GLES2_Renderer));
 }
 
 void IJK_GLES2_Renderer_free(IJK_GLES2_Renderer *renderer)
@@ -368,9 +371,11 @@ GLboolean IJK_GLES2_Renderer_renderOverlay(IJK_GLES2_Renderer *renderer, SDL_Vou
 
     glClear(GL_COLOR_BUFFER_BIT);               IJK_GLES2_checkError_TRACE("glClear");
 
+    GLsizei visible_width  = renderer->frame_width;
+    GLsizei visible_height = renderer->frame_height;
     if (overlay) {
-        GLsizei visible_width  = overlay->w;
-        GLsizei visible_height = overlay->h;
+        visible_width  = overlay->w;
+        visible_height = overlay->h;
         if (renderer->frame_width   != visible_width    ||
             renderer->frame_height  != visible_height   ||
             renderer->frame_sar_num != overlay->sar_num ||
@@ -380,36 +385,38 @@ GLboolean IJK_GLES2_Renderer_renderOverlay(IJK_GLES2_Renderer *renderer, SDL_Vou
             renderer->frame_height  = visible_height;
             renderer->frame_sar_num = overlay->sar_num;
             renderer->frame_sar_den = overlay->sar_den;
-            renderer->vertices_changed = 1;
         }
 
-        if (renderer->vertices_changed) {
-            renderer->vertices_changed = 0;
+        renderer->last_buffer_width = renderer->func_getBufferWidth(renderer, overlay);
 
-            IJK_GLES2_Renderer_Vertices_apply(renderer);
-            IJK_GLES2_Renderer_Vertices_reloadVertex(renderer);
-        }
-
-        GLsizei buffer_width   = renderer->func_getBufferWidth(renderer, overlay);
-        if (buffer_width > 0 &&
-            buffer_width > visible_width &&
-            buffer_width != renderer->buffer_width &&
-            visible_width != renderer->visible_width) {
-            renderer->buffer_width  = buffer_width;
-            renderer->visible_width = visible_width;
-
-            GLsizei padding_pixels     = buffer_width - visible_width;
-            GLfloat padding_normalized = ((GLfloat)padding_pixels) / buffer_width;
-            ALOGI("[yuv420p] padding changed: %d - %d = %d (%f)\n",
-                  buffer_width, visible_width,
-                  padding_pixels, padding_normalized);
-            IJK_GLES2_Renderer_TexCoords_reset(renderer);
-            IJK_GLES2_Renderer_TexCoords_cropRight(renderer, padding_normalized);
-            IJK_GLES2_Renderer_TexCoords_reloadVertex(renderer);
-        }
-        
         if (!renderer->func_uploadTexture(renderer, overlay))
             return GL_FALSE;
+    } else {
+        // NULL overlay means force reload vertice
+        renderer->vertices_changed = 1;
+    }
+
+    GLsizei buffer_width = renderer->last_buffer_width;
+    if (renderer->vertices_changed ||
+        (buffer_width > 0 &&
+         buffer_width > visible_width &&
+         buffer_width != renderer->buffer_width &&
+         visible_width != renderer->visible_width)){
+
+        renderer->vertices_changed = 0;
+
+        IJK_GLES2_Renderer_Vertices_apply(renderer);
+        IJK_GLES2_Renderer_Vertices_reloadVertex(renderer);
+
+        renderer->buffer_width  = buffer_width;
+        renderer->visible_width = visible_width;
+
+        GLsizei padding_pixels     = buffer_width - visible_width;
+        GLfloat padding_normalized = ((GLfloat)padding_pixels) / buffer_width;
+
+        IJK_GLES2_Renderer_TexCoords_reset(renderer);
+        IJK_GLES2_Renderer_TexCoords_cropRight(renderer, padding_normalized);
+        IJK_GLES2_Renderer_TexCoords_reloadVertex(renderer);
     }
 
     glDrawArrays(GL_TRIANGLE_STRIP, 0, 4);      IJK_GLES2_checkError_TRACE("glDrawArrays");