+++ init-android-exo.sh
@@ -17,7 +17,7 @@
 
 IJK_EXO_UPSTREAM=https://github.com/google/ExoPlayer.git
 IJK_EXO_FORK=https://github.com/google/ExoPlayer.git
-IJK_EXO_COMMIT=r1.5.2
+IJK_EXO_COMMIT=r1.5.8
 IJK_EXO_LOCAL_REPO=extra/ExoPlayer
 
 set -e
@@ -26,7 +26,7 @@ TOOLS=tools
 echo "== pull ExoPlayer base =="
 sh $TOOLS/pull-repo-base.sh $IJK_EXO_UPSTREAM $IJK_EXO_LOCAL_REPO
 
-echo "== pull libyuv fork =="
+echo "== pull ExoPlayer fork =="
 cd extra/ExoPlayer
 git checkout ${IJK_EXO_COMMIT} -B ijkplayer
 cd -
@@ -52,5 +52,5 @@ install_java player/ExtractorRendererBuilder.java
 install_java player/HlsRendererBuilder.java
 install_java player/SmoothStreamingRendererBuilder.java
 install_java EventLogger.java
-# install_java SmoothStreamingTestMediaDrmCallback.java
+install_java SmoothStreamingTestMediaDrmCallback.java
 # install_java WidevineTestMediaDrmCallback.java