+++ ios/IJKMediaDemo/IJKMediaDemo/IJKDemoSampleViewController.m
@@ -38,7 +38,7 @@
     [sampleList addObject:@[@"bipbop basic master playlist",
                             @"http://devimages.apple.com.edgekey.net/streaming/examples/bipbop_4x3/bipbop_4x3_variant.m3u8"]];
     [sampleList addObject:@[@"bipbop basic 400x300 @ 232 kbps",
-                            @"ijkhttphook://devimages.apple.com.edgekey.net/streaming/examples/bipbop_4x3/gear1/prog_index.m3u8"]];
+                            @"http://devimages.apple.com.edgekey.net/streaming/examples/bipbop_4x3/gear1/prog_index.m3u8"]];
     [sampleList addObject:@[@"bipbop basic 640x480 @ 650 kbps",
                             @"http://devimages.apple.com.edgekey.net/streaming/examples/bipbop_4x3/gear2/prog_index.m3u8"]];
     [sampleList addObject:@[@"bipbop basic 640x480 @ 1 Mbps",